FVDS: Safety wrapper

+++ fvds.sh
@@ -1,25 +1,31 @@
 #!/bin/bash
-if [ "$(grep -Ei 'debian|buntu|mint' /etc/*release)" ]; then
-	echo "Cleaning APT packages"
-	apt clean
-	sleep 2
-	echo "Done cleaning APT packages"
-	echo "Cleaning ~/.cache and ~/.local"
-	rm -rf ~/.cache ~/.local 
-	echo "Performing TRIM operations"
-	for mount in / /boot /home /swap; do
-		fstrim $mount;
-	done
-	echo "All done!"
-elif [ "$(grep -Ei 'fedora|redhat|rhel' /etc/*release)" ]; then
-	echo "Cleaning YUM/DNF packages"
-	yum clean all
-	sleep 2
-	dnf clean all
-	echo "Done cleaning YUM/DNF packages"
-	echo "Performing TRIM operations"
-	for mount in / /boot /home /swap; do
-		fstrim $mount;
-	done
-	echo "All done!"
+echo "This will delete data! (downloaded packages, the ~/.cache and ~/.local folders) - are you sure? [y/n]";
+read response;
+if [ $response != "y" ]; then
+	echo "Aborting.";
+else
+	if [ "$(grep -Ei 'debian|buntu|mint' /etc/*release)" ]; then
+		echo "Cleaning APT packages"
+		apt clean
+		sleep 2
+		echo "Done cleaning APT packages"
+		echo "Cleaning ~/.cache and ~/.local"
+		rm -rf ~/.cache ~/.local 
+		echo "Performing TRIM operations"
+		for mount in / /boot /home /swap; do
+			fstrim $mount;
+		done
+		echo "All done!"
+	elif [ "$(grep -Ei 'fedora|redhat|rhel' /etc/*release)" ]; then
+		echo "Cleaning YUM/DNF packages"
+		yum clean all
+		sleep 2
+		dnf clean all
+		echo "Done cleaning YUM/DNF packages"
+		echo "Performing TRIM operations"
+		for mount in / /boot /home /swap; do
+			fstrim $mount;
+		done
+		echo "All done!"
+	fi
 fi
